fix(cli): add check for required files
After failing to deploy the sunpath app I realized that we probably want to make sure the DockerFile and app.py exist in the folder before scheduling it for deployment.

Also added some extra note to docstring and printed out some extra messages to let the user know where to check the app after the deployment.